Move BannerListener from GatewayApi to JobScheduler for AB#13736
