+++ Apps/Common/test/unit/Utils/Scenario.cs
@@ -13,27 +13,31 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 // -------------------------------------------------------------------------
-namespace HealthGateway.GatewayApi.Converters
+namespace HealthGateway.CommonTests.Utils
 {
-    using System;
-    using System.Text.Json;
-    using System.Text.Json.Serialization;
-
     /// <summary>
-    /// Json Converter to put Postgres +00 dates into c# UTC DateTimes.
+    /// Scenario enum to use for testing purposes.
     /// </summary>
-    public class DateTimeConverter : JsonConverter<DateTime>
+    public enum Scenario
     {
-        /// <inheritdoc/>
-        public override DateTime Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-        {
-            return reader.GetDateTime().ToUniversalTime();
-        }
+        /// <summary>
+        /// The communication is Active.
+        /// </summary>
+        Active,
 
-        /// <inheritdoc/>
-        public override void Write(Utf8JsonWriter writer, DateTime value, JsonSerializerOptions options)
-        {
-            writer.WriteStringValue(value.ToUniversalTime());
-        }
+        /// <summary>
+        /// The communication is expired.
+        /// </summary>
+        Expired,
+
+        /// <summary>
+        /// The communication should be removed.
+        /// </summary>
+        Deleted,
+
+        /// <summary>
+        /// The communication should be future effective dated.
+        /// </summary>
+        Future,
     }
 }
